refactor(datasets): expose PREVIEW_FEATURE_CAP through the domain service façade
PREVIEW_FEATURE_CAP is defined in service_analysis.py but was missing
from the domain/service.py façade's re-exports, forcing three test
files to import the sub-module directly against the façade rule
enforced by test_no_external_imports_of_dataset_domain_submodules.

Adds it to the façade's imports and __all__, and switches the three
call sites to import from the façade. test_analysis_curved_sources.py
(private-symbol regression pin) and the two sites that patch the
submodule directly (test_vrt_delete_guard_174.py,
test_workflow_extension.py) are unchanged — they need the submodule
path, not the façade.
