+++ backend/tests/test_analysis_spatial_join.py
@@ -17,7 +17,7 @@ from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.modules.catalog.datasets.api import router_analysis
-from app.modules.catalog.datasets.domain.service_analysis import PREVIEW_FEATURE_CAP
+from app.modules.catalog.datasets.domain.service import PREVIEW_FEATURE_CAP
 from app.platform.analysis_sql import MAX_SOURCE_FEATURES
 from app.processing.analysis.tasks import _materialize
 from app.processing.ingest.metadata import add_4326_column
